feat(form-builder): Extract asset select handler into shared file
